dialog

+++ app/src/main/java/br/com/raspemania/model/entidade/PremiacaoList.java
@@ -3,6 +3,6 @@ package br.com.raspemania.model.entidade;
 public class PremiacaoList {
 
     public int quantidadePremiada;
-    public long valorPremiado;
+    public float valorPremiado;
 
 }
